[stats] statfuncs: Wconversion fixes

+++ include/sserialize/stats/statfuncs.h
@@ -72,17 +72,17 @@ void linearRegression(TIterator begin, const TIterator & end, double & slope, do
 	for(; begin != end; ++begin, ++size) {
 		cysum += *begin;
 	}
-	double meanY = (double)(cysum/size) + static_cast<double>(cysum%size)/size;
+	double meanY = (double)(cysum/size) + static_cast<double>(cysum%size)/static_cast<double>(size);
 	double meanX = (double)(size-1)/2.0;
 	double count = 0.0;
 	slope = 0.0;
 	for(TIterator it(begin); it != end; ++it, count += 1.0) {
 		slope += (count - meanX)*(*it-meanY);
 	}
-	slope /= size-1; //(n)
+	slope /= (double)(size-1); //(n)
 	slope *= 12;
-	slope /= size+1; //(n+2)
-	slope /= size; //(n+1)
+	slope /= (double)(size+1); //(n+2)
+	slope /= (double)size; //(n+1)
 	yintercept = meanY - slope*meanX;
 }
